seperated functions into a file
